feat: add mcp subcommand to CLI to run Codex as an MCP server
